fix display things
