+++ software/apps/romi_control/helpers.c
@@ -1,6 +1,72 @@
 #include "helpers.h"
 
-NRF_TWI_MNGR_DEF(twi_mngr_instance, 5, 0);
+//NRF_TWI_MNGR_DEF(twi_mngr_instance, 5, 0);
+
+
+// Global Variables 
+
+KobukiSensors_t sensors = {0};
+ret_code_t error_code = NRF_SUCCESS;
+
+
+// Intervals for advertising and connections
+simple_ble_config_t ble_config = {
+        // c0:98:e5:49:xx:xx
+        .platform_id       = 0x46,    // used as 4th octect in device BLE address
+        .device_id         = 0xEEC5, 
+        .adv_name          = "Robo-AR", // used in advertisements if there is room
+        .adv_interval      = MSEC_TO_UNITS(1000, UNIT_0_625_MS),
+        .min_conn_interval = MSEC_TO_UNITS(100, UNIT_1_25_MS),
+        .max_conn_interval = MSEC_TO_UNITS(200, UNIT_1_25_MS),
+};
+
+//4607eda0-f65e-4d59-a9ff-84420d87a4ca
+simple_ble_service_t robot_service = {{
+    .uuid128 = {0xca,0xa4,0x87,0x0d,0x42,0x84,0xff,0xA9,
+                0x59,0x4D,0x5e,0xf6,0xa0,0xed,0x07,0x46}
+}};
+
+//Declare characteristics and variables for your service
+simple_ble_char_t waypoint_char = {.uuid16 = 0xeda1};
+simple_ble_char_t ack_char = {.uuid16 = 0xeda2};
+simple_ble_app_t* simple_ble_app;
+
+
+
+const float CONVERSION = 0.0006108;
+const float angle_threshold = .5;
+const float distance_threshold = .02;
+int k_dist = 190;
+int k_diff = 250;
+
+
+float waypoint[2] = {0, 0};
+bool new_waypoint_written = false;
+int acknowledged = -1;
+bool connected = false;
+
+
+void ble_evt_connected(ble_evt_t const* p_ble_evt) {
+    connected = true;
+}
+
+void ble_evt_disconnected(ble_evt_t const* p_ble_evt) {
+    connected = false;
+}
+
+// void read_input() {
+//     printf("Bluetooth message recieved\n");
+//     if (state == WAITING) {
+//         printf("Distance: %f\n", waypoint[0]);
+//         printf("Angle: %f\n", waypoint[1]);
+//     }
+// }
+
+void ble_evt_write(ble_evt_t const* p_ble_evt) {
+    //logic for each characteristic and related state changes
+    //Try not to modify the state here...
+    new_waypoint_written = true;
+}
 
 void setup() {
   // initialize RTT library
@@ -34,40 +100,9 @@ void setup() {
   nrf_gpio_pin_set(23);
   nrf_gpio_pin_set(24);
   nrf_gpio_pin_set(25);
+}
 
-
-
-  // initialize display
-  nrf_drv_spi_t spi_instance = NRF_DRV_SPI_INSTANCE(1);
-  nrf_drv_spi_config_t spi_config = {
-    .sck_pin = BUCKLER_LCD_SCLK,
-    .mosi_pin = BUCKLER_LCD_MOSI,
-    .miso_pin = BUCKLER_LCD_MISO,
-    .ss_pin = BUCKLER_LCD_CS,
-    .irq_priority = NRFX_SPI_DEFAULT_CONFIG_IRQ_PRIORITY,
-    .orc = 0,
-    .frequency = NRF_DRV_SPI_FREQ_4M,
-    .mode = NRF_DRV_SPI_MODE_2,
-    .bit_order = NRF_DRV_SPI_BIT_ORDER_MSB_FIRST
-  };
-  error_code = nrf_drv_spi_init(&spi_instance, &spi_config, NULL, NULL);
-  APP_ERROR_CHECK(error_code);
-  display_init(&spi_instance);
-  printf("Display initialized!\n");
-
-  // initialize i2c master (two wire interface)
-  nrf_drv_twi_config_t i2c_config = NRF_DRV_TWI_DEFAULT_CONFIG;
-  i2c_config.scl = BUCKLER_SENSORS_SCL;
-  i2c_config.sda = BUCKLER_SENSORS_SDA;
-  i2c_config.frequency = NRF_TWIM_FREQ_100K;
-  error_code = nrf_twi_mngr_init(&twi_mngr_instance, &i2c_config);
-  APP_ERROR_CHECK(error_code);
-  lsm9ds1_init(&twi_mngr_instance);
-  printf("IMU initialized!\n");
-
-  // initialize Kobuki
-  kobukiInit();
-  printf("Kobuki initialized!\n");
-	
-
+float measure_distance(uint16_t current_encoder, uint16_t previous_encoder) {
+   float distance = CONVERSION * (current_encoder - previous_encoder);
+   return distance;
 }
