+++ software/apps/romi_control/main.c
@@ -298,15 +298,49 @@
 // }
 
 #include "states.h"
-#include "outputs.h"
-#include "helpers.h"
-#include "inputs.h"
+NRF_TWI_MNGR_DEF(twi_mngr_instance, 5, 0);
+
 
 int main(void) {
+  // initialize display
+  nrf_drv_spi_t spi_instance = NRF_DRV_SPI_INSTANCE(1);
+  nrf_drv_spi_config_t spi_config = {
+    .sck_pin = BUCKLER_LCD_SCLK,
+    .mosi_pin = BUCKLER_LCD_MOSI,
+    .miso_pin = BUCKLER_LCD_MISO,
+    .ss_pin = BUCKLER_LCD_CS,
+    .irq_priority = NRFX_SPI_DEFAULT_CONFIG_IRQ_PRIORITY,
+    .orc = 0,
+    .frequency = NRF_DRV_SPI_FREQ_4M,
+    .mode = NRF_DRV_SPI_MODE_2,
+    .bit_order = NRF_DRV_SPI_BIT_ORDER_MSB_FIRST
+  };
+  error_code = nrf_drv_spi_init(&spi_instance, &spi_config, NULL, NULL);
+  APP_ERROR_CHECK(error_code);
+  display_init(&spi_instance);
+  printf("Display initialized!\n");
+
+    // initialize i2c master (two wire interface)
+  nrf_drv_twi_config_t i2c_config = NRF_DRV_TWI_DEFAULT_CONFIG;
+  i2c_config.scl = BUCKLER_SENSORS_SCL;
+  i2c_config.sda = BUCKLER_SENSORS_SDA;
+  i2c_config.frequency = NRF_TWIM_FREQ_100K;
+  error_code = nrf_twi_mngr_init(&twi_mngr_instance, &i2c_config);
+  APP_ERROR_CHECK(error_code);
+  lsm9ds1_init(&twi_mngr_instance);
+  printf("IMU initialized!\n");
+
+  // initialize Kobuki
+  kobukiInit();
+  printf("Kobuki initialized!\n");
+
   setup();
   system_state_t system_state = init_state();
   while(true) {
     inputs_t inputs = get_inputs();
-    transition(inputs, &system_state);
+    outputs_t output = transition(inputs, &system_state);
+    display_write(output.display_line_0, DISPLAY_LINE_0);
+    display_write(output.display_line_1, DISPLAY_LINE_1);
+    do_outputs(output);
   }
 }
