+++ software/apps/romi_control/outputs.c
@@ -6,6 +6,4 @@ void do_outputs(outputs_t output) {
 	if (output.notify_ack >= 0) {
 		simple_ble_notify_char(&ack_char);
 	}
-	display_write(output.display_line_0, DISPLAY_LINE_0);
-	display_write(output.display_line_1, DISPLAY_LINE_1);
 }
